requirements.txt
fastmcp>=2.8,<3.0
# TA-Lib-Precompiled>=0.4.0  # Uncomment when TA-Lib works with Python 3.13
numpy<2.0
python-dotenv>=1.0.0
pydantic>=2.0
pytest>=8.0
pytest-asyncio>=0.23
flake8>=7.0
black>=24.4.0